requirements.txt
Flask==2.0.2
tensorflow==2.7.0
Pillow==9.0.0
